refactor: @.type.ts 형태로 types 폴더에 있는 파일들의 이름을 통일
