+++ backend/src/journals/types/create-journal-data.type.ts
