+++ backend/src/journals/types/journal-detail.type.ts
