+++ backend/src/journals/types/journal-info.type.ts
